+++ lib/diaspora/camo.rb
@@ -17,7 +17,11 @@ module Diaspora
       return unless url
       return url unless self.url_eligible?(url)
 
-      url = Addressable::URI.encode(Addressable::URI.unencode(url))
+      begin
+        url = Addressable::URI.encode(Addressable::URI.unencode(url))
+      rescue Addressable::URI::InvalidURIError
+        return url
+      end
 
       digest = OpenSSL::HMAC.hexdigest(
         OpenSSL::Digest.new('sha1'),