+++ Changelog.md
@@ -1,15 +1,14 @@
 # 0.7.8.0
 
 ## Refactor
-
 * Make setting up a development environment 9001% easier by adding a Docker-based setup [#7870](https://github.com/diaspora/diaspora/pull/7870)
 * Improve `web+diaspora://` handler description [#7909](https://github.com/diaspora/diaspora/pull/7909)
 * Move comment timestamp next to author name [#7905](https://github.com/diaspora/diaspora/pull/7905)
 
 ## Bug fixes
+* Ignore invalid URLs for camo [#7922](https://github.com/diaspora/diaspora/pull/7922)
 
 ## Features
-
 * Add the ability to assign roles in the admin panel [#7868](https://github.com/diaspora/diaspora/pull/7868)
 * Improve memory usage with libjemalloc if available [#7919](https://github.com/diaspora/diaspora/pull/7919)